examples: test_pps_input: Also test for ref_locked on E31x
The E31x will always have a valid PPS, because it is generated
internally. The external PPS however is used to drive the ppsloop, which
means the ref_locked sensor can be a substitute for the PPS, but only on
in this case.

This commit also removes some superfluous includes from this example.