muxed_zero_copy_if: Add timeout parameter
The muxed_zero_copy_if used a timeout of zero for fetching the next
packet from its inbound transport ("receive"). This can rail CPUs
depending on the transport used.

This adds a timeout parameter that is applied to the recv_buff call in
the demuxer thread. Many demuxers don't need small timeouts (e.g.
control) because they idle most of the time.