mpm: gpsd_iface: Sanity check calls with timeouts
All the queries in gpsd_iface assume that if the device is responsive,
it will eventually start producing valid data. However, cases have been
observed (on the E310) where that is not the case, and the GPS mode can
remain on 0 in the TPV result. This would lead to an infinite loop.

The solution is to add (long) timeouts to all calls, which will throw
an exception after the default value of 60 seconds.

This applies to all sensors calls (get_gps_tpv_sensor etc.).