mpm: Fix full-duplex streaming for muxed xports
On E310, the data streams are muxed (theoretically, other devices could
also mux but don't). There was a bug where the container for the
muxed_zero_copy_if was a scalar, but we need one muxed_zero_copy_if per
DMA channel (so, at least two: One for RX, one for TX).
Without this fix, full-duplex would always fail on the E310, typically
with timeouts on RX.