Recent versions of the GCC compiler suite have become much more inclined to warn of otherwise benign code practices and forms. Notable are switch fallthrough cases, and unused functions or variables.
But the compiler also warns about operator= overloads that are already
addressed by constructors.

F'rinstance,  consider a type Foo that can be built from a std::string
so it has a constructor
  Foo(std::string v)
If Foo also defines
  Foo operator=(std::string v)
then the two should be doing exactly the same thing.  Further, the
compiler can't figure out which it should use when it sees this:
  Foo foo = std::string("foostring");
Should it use the assignment operator?  or should it promote (?)
the std::string up to a Foo using the implicit type conversion provided
by the constructor?

Addressing the many warnings takes very few code changes, none of which
affect the behavior of the code.  One should consider though that the
warnings are not just noise.  Real bugs can be burried under the warnings,
and the more that appear in the log, the greater the chances of missing
one that "really matters."

changes:
-------------
include/uhd/types/sid.hpp

Removed operator= definiitons for uint32_t, sid_t, and std::string to sid_t.
These are already provided by the converting constructors.

-------------
lib/deps/rpclib/include/rpc/msgpack/object.hpp
host/lib/transport/super_recv_packet_handler.hpp

The memcpy calls get flagged for doing a blind naive copy of one object type
to another.  This is really a rather gross thing to do, but apparently was
the intent of the developer.  If that's the way one wants to hang oneself,
then the proper form is to coerce each of the pointers to a void* as
the type information is being discarded and the writer wishes it to be
ignored.

super_recv_packet_handler has a call to memset that offends gcc in a similar
manner.

I wouldn't do it that way.

-------------
host/lib/transport/libusb1_base.cpp

The libusb_set_debug function was deprecated some time ago.  However, there
may be platforms that are far enough behind that the code should check the
API version and use the deprecated form for older versions and the new
lib_set_option form for newer versions.  Developers will need to decide how
old the libusb library can be before it is unsupported.  The CMakeLists.txt
files do not apparently place a version requirement on libusb.

-------------
host/lib/transport/uhd-dpdk/uhd_dpdk.c

The compiler frets over the 0x%0llx format specifier as it wants to see
a long-long rather than a long argument.

-------------
host/lib/transport/uhd-dpdk/uhd_dpdk_driver.c

The compiler complains about an uninitialized use of sock_req.  From what
I can tell, the code is perfectly valid and does all the right things.
However, it only takes assignment of a default NULL value to appease the
persnickety compiler. So, "struct uhd_dpdk_config_req * sock_req;"
becomes "struct uhd_dpdk_config_req * sock_req = NULL;"

The compiler (!) also complains about snprintf (does it occur to you that
the compiler is poking into things a little deep?)  producing a string
that may be too big for the buffer, and that some characters might be lost.
The problem is the 16 character name, a limit imposed by pthreads and the
rest of the stack.  The problem is solved by limiting the width of the %u
specifier like this "%3hhu" -- the hh considers the corresponding argument
as a char...(sigh).

-------------
host/lib/transport/uhd-dpdk/uhd_dpdk_fops.c

Another problem with the use of a possibly (though not actually) un-initialized
pointer.  Fixed by splitting the declaration from the first assignment. (wow.)

-------------
host/lib/usrp/dboard/rhodium/rhodium_radio_ctrl_cpld.cpp

There are two "unused functions" here in an anonymous namespace.
Perhaps they should be removed, but on the off chance that someone
cares, the unused-function warning is turned off locally.  (The warning
is useful in steering developers to mis-speled (sic) function names,
perhaps.  It never did me any good.)

-------------
host/lib/usrp/dboard/rhodium/rhodium_radio_ctrl_impl.cpp

The set_rx_bandwidth and set_tx_bandwidth methods both leave
the "bandwidth" argument unused unless UHD_LOG_TRACE is defined
as something that actually *does* something.  Interesting, and perhaps
an indication of a mis-feature.

-------------
host/lib/usrp/mpmd/mpmd_xport_ctrl_dpdk_udp.cpp
host/lib/usrp/x300/x300_impl.cpp

The dpdk_zero_copy::make method takes a dpdk_port_id parameter.
It is "coerced" into an unsigned int here, but the code also
included a "const" prefix.  The compiler warns about it, as
the notion of telling the callee that a parameter is const, is
not the prerogative of the caller.  The "const" has been removed,
and the code checked to ensure that the "dpdk_port_id parameter
is not used *after* the call to make.

-------------
host/lib/usrp/usrp1/soft_time_ctrl.cpp

In a prime illustration of just the kind of dumpster fire the
recent obsession with warning about every possible manifestation
of the "goto fail" bug, gcc is now warning about fallthrough case
sequences.  C++17 defines a marker "[[fallthrough]]" to signal that
the fallthrough behavior is intended.  This isn't necessarily
implemented by other compilers.  The gnu compiler uses [[gnu::fallthrough]],
and that's been applied here.  Note that this has not been tested
against other compilers.  It may be that the best choice is to define
a macro that depends on the compiler version and such.

-------------
host/lib/usrp/x300/cdecode.c

Again with the fallthrough stuff.  But in this case, it is in a C file,
so we use the standard "__attribute__((fallthrough))"

There, doesn't that feel better?

-------------
host/tests/sid_t_test.cpp

In order for us to get the assignment right, the "char*" right-hand-sides
needed to be converted to std::string.  Perhaps that means that the sid_t
needs a constructor that takes a char*.