Changes to eliminate compiler warnings.  Recent versions of gcc can actually generate bad code in the presence of behavior that triggers what looks to be a warning.
All warnings were encountered with gcc 9.2.1

host/lib/deps/rpclib/include/rpc/msgpack/object.hpp
Compiler complains (and has complained for some time) about a blind
memcpy on two objects that may or may not be compatible. This behavior
is risky, but that's the way the implementers apparently wanted it.
So, the right way to do this is to cast the pointers to void and silence
the warning, though we wish one could do better.

host/lib/rfnoc/mgmt_portal.cpp
A variable "node_addr" was declared and initialized for later use in a
UHD_LOG_DEBUG statement.  As the existence of the call is conditioned on
configuration settings, when UHD_LOG_DEBUG is not expanded, the node_addr
value is unused.  The compiler complains.  Examination indicates that this
issue is benign. (Compilers warn of this because the condition may indicate
a possible mis-spelling that is syntactically irrelevant, but may be exposing
a value in an outer scope with a similar name.  Paranoid? Yup. Addresses a real
issue?  Yup.)
Providing an empty statement that treats node_addr as a void suffices to quiet
the warning, and is an accpetable and widely used pattern.

host/lib/rfnoc/rfnoc_graph.cpp
A function definition declares a parameter that is only used when UHD_LOG_TRACE
is expanded.  See above.  (void) mb_idx; fixes the problem.

host/lib/transport/libusb1_base.cpp
libusb folks changed the api to set debug level.  It once was libusb_set_debug
but now prefers lib_usb_set_option(.... )  The compiler complains that the
old libusb_set_debug routine is now deprecated.
This is now conditioned on LIBUSB_API_VERSION and wrapped in a UHD_USB_SET_DEBUG macro
in libuxb1_base.cpp

host/lib/usrp/common/ad9361_driver/ad9361_device.cpp
unused parameter "rate" only used when UHD_LOG_TRACE is expanded.  (void) rate satisfies
the compiler.

host/lib/usrp/dboard/rhodium/rhodium_radio_control_cpld.cpp
Compiler complains of unused functions "rx_band_to_log" and "tx_band_to_log".
The only real way to quiet the warning from gcc is with a pragma for GCC diagnostics.
As done here.

host/tests/client_zero_test.cpp
This is an interesting one: The original code was
        if ((addr & SLOT_OFFSET) == 1) {
But SLOT_OFFSET is 512.  This statement is *always* true.  I'm assuming that the
intent was to test to see that addr has the SLOT_OFFSET bit set.  Then:
        if ((addr & SLOT_OFFSET) != 0) {
is likely the answer.   This conditional triggers a call to UHD_LOG_INFO, so
it may not be covered by any test, and is unlikely to cause an operational issue.