mpm: cmake: adopt default enable values of modules
The cmake script (macro MPM_REGISTER_COMPONENT) has an issue that the default
value is ignored which will be fixed in the following commit.

As a result of this issue, the modules Mykonos, Magnesium, E320 and E300 never
got enabled per default. It was always necessary to use an override
like ENABLE_MYKONOS=ON.

Therefore, change the default enable values so that a fix to MPM_REGISTER_COMPONENT
does not change the current behavior.