lib: deps: Add FlatBuffers 1.11.0 header files
FlatBuffers is used to (de-)serialize calibration data. For most use
cases, only a few headers are required, so we provide them with UHD
instead of requiring all UHD users to install their own version of
FlatBuffers.

This adds the headers to host/lib/deps, and updates the Debian copyright
file appropriately.