cmake: Add CMakeRC module and ::rc namespace
This module allows to add binary files into UHD (e.g., for cal data
storage) in a platform-independent way.

The upstream CMakeRC.cmake is modified in the following way to allow
linkage against a shared object:

```diff
     # Generate the actual static library. Each source file is just a single file
     # with a character array compiled in containing the contents of the
     # corresponding resource file.
     add_library(${name} STATIC ${libcpp})
     set_property(TARGET ${name} PROPERTY CMRC_LIBDIR "${libdir}")
     set_property(TARGET ${name} PROPERTY CMRC_NAMESPACE "${ARG_NAMESPACE}")
     target_link_libraries(${name} PUBLIC cmrc::base)
     set_property(TARGET ${name} PROPERTY CMRC_IS_RESOURCE_LIBRARY TRUE)
+    set_property(TARGET ${name} PROPERTY POSITION_INDEPENDENT_CODE ON)
```

This forces the -fPIC flag for the static object that CMakeRC generates,
allowing to link it into a shared object file.

The version of CMakeRC used is: a7e355290, cloned from
git@github.com:vector-of-bool/cmrc.git.