utils: cal: Replace property tree accesses with other API calls
The cal utils read the property tree to identify the daughterboard
information (serial, type, etc.). The paths for properties have changed
in UHD 4.0, causing these utilities to fail.

Instead of conditionally parsing the property tree, this commit changes
the utilities such that dictionaries are pulled out of the multi_usrp
object, which work regardless of the device type.