utils: uhd_images_downloader: use HTTPS instead of HTTP to avoid redirect
Since the webserver responding to requests at http://files.ettus.com
replies with a redirect to the same URL, but using the https://
transport, this isn't only the safer thing we should be doing, anyways,
but also the quicker, since it saves the user from one unnecessary
redirect.

This is the 3.15-LTS variant of this fix.