cores: Fix rx_vita_core_3000 assertion error in issue_stream_cmd()
The current code had an assertion

    UHD_ASSERT_THROW(stream_cmd.num_samps <= 0x0fffffff);

which would check that num_samps in a stream command don't exceed the
counter depth in the FPGA. However, this is only relevant if the stream
command is not "continuous" or "stop".
num_samps could be unitialized, and randomly have a value larger than
the maximum, and the assertion could trigger even though the value in
num_samps is irrelevant.

The new assertion checks for the correct case, and has a more verbose
error message.